+++ common-model-repository/src/main/java/cn/featherfly/common/repository/AliasField.java
@@ -1,0 +1,9 @@
+package cn.featherfly.common.repository;
+
+/**
+ * alias field.
+ *
+ * @author zhongj
+ */
+public interface AliasField extends Field, Alias {
+}